+++ source/common/http/hash_policy.cc
@@ -112,9 +112,8 @@ private:
 
 HashPolicyImpl::HashPolicyImpl(
     absl::Span<const envoy::config::route::v3alpha::RouteAction::HashPolicy* const> hash_policies) {
-  // TODO(htuch): Add support for cookie hash policies, #1295
-  hash_impls_.reserve(hash_policies.size());
 
+  hash_impls_.reserve(hash_policies.size());
   for (auto* hash_policy : hash_policies) {
     switch (hash_policy->policy_specifier_case()) {
     case envoy::config::route::v3alpha::RouteAction::HashPolicy::PolicySpecifierCase::kHeader: