+++ source/extensions/filters/network/dubbo_proxy/filters/filter_config.h
@@ -38,7 +38,7 @@ public:
   createFilterFactoryFromProto(const Protobuf::Message& config, const std::string& stat_prefix,
                                Server::Configuration::FactoryContext& context) PURE;
 
-  std::string category() const override { return "dubbo_proxy.filters"; }
+  std::string category() const override { return "envoy.dubbo_proxy.filters"; }
 };
 
 } // namespace DubboFilters